[Comp-843] IR docx formatting changes
