+++ compliance-api/src/compliance_api/services/inspection_record/ir_template_constant.py
@@ -3,6 +3,7 @@
 INSPECTION_SCOPE = """<p class="editor-paragraph" dir="ltr">
     <span>The Officer inspected {{ area_inspected }}.</span>
 </p>
+<p class="editor-paragraph">&nbsp;</p>
 <p class="editor-paragraph" dir="ltr">
     <span>The inspection included a debrief of observations with Project staff on {{ debrief_date }}.</span>
 </p>
