+++ compliance-api/tests/unit/services/inspection_record/test_inspection_record_doc_generator.py
@@ -4,13 +4,13 @@ from bs4 import BeautifulSoup
 from docx import Document
 from docx.shared import Pt
 
-from compliance_api.services.inspection_record.inspection_record_doc_generator import (
-    _add_formatted_text_to_para, _add_formatted_text_to_table_para, _add_html_paragraphs_to_cell,
-    _add_html_to_container, _add_paragraph)
+from compliance_api.services.inspection_record.html_to_docx import (
+    _add_formatted_text_to_container, _add_formatted_text_to_table_cell, _add_html_to_container,
+    _add_html_paragraphs_to_cell, _add_paragraph)
 
 
-class TestAddFormattedTextToPara:
-    """Tests for _add_formatted_text_to_para function."""
+class TestAddFormattedTextToContainer:
+    """Tests for _add_formatted_text_to_container function."""
 
     def _create_document_with_paragraph(self):
         """Create a document with a paragraph for testing."""
@@ -20,25 +20,25 @@ class TestAddFormattedTextToPara:
 
     def test_plain_text_is_added(self):
         """Test that plain text is correctly added to paragraph."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Plain text content</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         assert para.text == "Plain text content"
 
     def test_bold_text_is_formatted(self):
         """Test that bold text has bold formatting applied."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Text with <strong>bold</strong> word</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         # Check that bold run exists and is marked bold
         bold_runs = [run for run in para.runs if run.bold]
@@ -47,13 +47,13 @@ class TestAddFormattedTextToPara:
 
     def test_italic_text_is_formatted(self):
         """Test that italic text has italic formatting applied."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Text with <em>italic</em> word</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         # Check that italic run exists and is marked italic
         italic_runs = [run for run in para.runs if run.italic]
@@ -62,78 +62,78 @@ class TestAddFormattedTextToPara:
 
     def test_space_preserved_before_bold(self):
         """Test that space before bold text is preserved."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Before <strong>bold</strong> after</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         full_text = para.text
         assert full_text == "Before bold after"
 
     def test_space_preserved_after_bold(self):
         """Test that space after bold text is preserved."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Before <strong>bold</strong> after</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         full_text = para.text
         assert full_text == "Before bold after"
 
     def test_space_preserved_before_italic(self):
         """Test that space before italic text is preserved."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Before <em>italic</em> after</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         full_text = para.text
         assert full_text == "Before italic after"
 
     def test_multiple_formatted_words_preserve_spaces(self):
         """Test that multiple formatted words preserve all spaces."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Text <strong>bold</strong> and <em>italic</em> words</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         full_text = para.text
         assert full_text == "Text bold and italic words"
 
     def test_consecutive_formatted_elements_preserve_spaces(self):
         """Test that consecutive formatted elements preserve spaces between them."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p><strong>Bold</strong> <em>italic</em></p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         full_text = para.text
         assert full_text == "Bold italic"
 
     def test_b_tag_treated_as_bold(self):
         """Test that <b> tag is treated the same as <strong>."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Text with <b>bold</b> word</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         bold_runs = [run for run in para.runs if run.bold]
         assert len(bold_runs) == 1
@@ -141,13 +141,13 @@ class TestAddFormattedTextToPara:
 
     def test_i_tag_treated_as_italic(self):
         """Test that <i> tag is treated the same as <em>."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         html = "<p>Text with <i>italic</i> word</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         italic_runs = [run for run in para.runs if run.italic]
         assert len(italic_runs) == 1
@@ -155,22 +155,22 @@ class TestAddFormattedTextToPara:
 
     def test_whitespace_only_text_node_preserved(self):
         """Test that whitespace-only text nodes between elements are preserved."""
-        _, para = self._create_document_with_paragraph()
+        doc, para = self._create_document_with_paragraph()
 
         # This HTML has a whitespace-only text node between </strong> and <em>
         html = "<p><strong>Bold</strong> <em>italic</em></p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_para(para, p_element, Pt(11))
+        _add_formatted_text_to_container(doc, p_element, Pt(11), first_para=para)
 
         # The space between Bold and italic must be preserved
         assert para.text == "Bold italic"
         assert " " in para.text
 
 
-class TestAddFormattedTextToTablePara:
-    """Tests for _add_formatted_text_to_table_para function."""
+class TestAddFormattedTextToTableCell:
+    """Tests for _add_formatted_text_to_table_cell function."""
 
     def _create_document_with_table_cell_paragraph(self):
         """Create a document with a table cell paragraph for testing."""
@@ -178,68 +178,68 @@ class TestAddFormattedTextToTablePara:
         table = doc.add_table(rows=1, cols=1)
         cell = table.rows[0].cells[0]
         para = cell.paragraphs[0]
-        return doc, para
+        return doc, cell, para
 
     def test_plain_text_is_added(self):
         """Test that plain text is correctly added to table paragraph."""
-        _, para = self._create_document_with_table_cell_paragraph()
+        _, cell, para = self._create_document_with_table_cell_paragraph()
 
         html = "<p>Plain text content</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_table_para(para, p_element)
+        _add_formatted_text_to_table_cell(cell, p_element, first_para=para)
 
         assert para.text == "Plain text content"
 
     def test_space_preserved_before_bold(self):
         """Test that space before bold text is preserved in table cells."""
-        _, para = self._create_document_with_table_cell_paragraph()
+        _, cell, para = self._create_document_with_table_cell_paragraph()
 
         html = "<p>Before <strong>bold</strong> after</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_table_para(para, p_element)
+        _add_formatted_text_to_table_cell(cell, p_element, first_para=para)
 
         full_text = para.text
         assert full_text == "Before bold after"
 
     def test_space_preserved_after_italic(self):
         """Test that space after italic text is preserved in table cells."""
-        _, para = self._create_document_with_table_cell_paragraph()
+        _, cell, para = self._create_document_with_table_cell_paragraph()
 
         html = "<p>Before <em>italic</em> after</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_table_para(para, p_element)
+        _add_formatted_text_to_table_cell(cell, p_element, first_para=para)
 
         full_text = para.text
         assert full_text == "Before italic after"
 
     def test_span_preserves_spaces(self):
         """Test that span elements preserve surrounding spaces."""
-        _, para = self._create_document_with_table_cell_paragraph()
+        _, cell, para = self._create_document_with_table_cell_paragraph()
 
         html = "<p>Before <span>span</span> after</p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_table_para(para, p_element)
+        _add_formatted_text_to_table_cell(cell, p_element, first_para=para)
 
         full_text = para.text
         assert full_text == "Before span after"
 
     def test_whitespace_only_text_node_preserved(self):
         """Test that whitespace-only text nodes between elements are preserved."""
-        _, para = self._create_document_with_table_cell_paragraph()
+        _, cell, para = self._create_document_with_table_cell_paragraph()
 
         html = "<p><strong>Bold</strong> <em>italic</em></p>"
         soup = BeautifulSoup(html, "html.parser")
         p_element = soup.find("p")
 
-        _add_formatted_text_to_table_para(para, p_element)
+        _add_formatted_text_to_table_cell(cell, p_element, first_para=para)
 
         assert para.text == "Bold italic"
